# backend/smart-space-booking/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register/member": {
      "post": {
        "operationId": "AuthController_registerMember",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Registrasi akun member / pelanggan baru",
        "tags": [
          "Autentikasi"
        ]
      }
    },
    "/api/auth/register/admin-space": {
      "post": {
        "operationId": "AuthController_registerAdminSpace",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAdminSpaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Registrasi pengelola lokasi coworking space",
        "tags": [
          "Autentikasi"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Login akun member atau admin space",
        "tags": [
          "Autentikasi"
        ]
      }
    },
    "/api/auth/profile": {
      "get": {
        "operationId": "AuthController_profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Profil dan hak akses pengguna yang sedang login",
        "tags": [
          "Autentikasi"
        ]
      }
    },
...
    "/api/admin/spaces": {
      "get": {
        "operationId": "AdminSpacesController_daftar",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Daftar seluruh space milik admin",
        "tags": [
          "Admin - Space"
        ]
      },
      "post": {
        "operationId": "AdminSpacesController_buat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Tambah space baru beserta fasilitas dan foto",
        "tags": [
          "Admin - Space"
        ]
      }
    },
    "/api/admin/spaces/{id}": {
      "get": {
        "operationId": "AdminSpacesController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Detail space berdasarkan ID",
        "tags": [
          "Admin - Space"
        ]
      },
      "put": {
        "operationId": "AdminSpacesController_perbarui",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Perbarui data ruangan dan fasilitas space",
        "tags": [
          "Admin - Space"
        ]
      },
      "delete": {
        "operationId": "AdminSpacesController_hapus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Hapus space (soft delete)",
        "tags": [
          "Admin - Space"
        ]
      }
    },
...
    "/api/admin/diskon/{id}": {
      "get": {
        "operationId": "AdminDiskonController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Detail kode promo berdasarkan ID",
        "tags": [
          "Admin - Kode Promo"
        ]
      },
      "put": {
        "operationId": "AdminDiskonController_perbarui",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDiskonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Perbarui kode promo dan periode berlakunya",
        "tags": [
          "Admin - Kode Promo"
        ]
      },
      "delete": {
        "operationId": "AdminDiskonController_hapus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Hapus kode promo (soft delete)",
        "tags": [
          "Admin - Kode Promo"
        ]
      }
    },
...
    "/api/spaces/{id}": {
      "get": {
        "operationId": "SpacesController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Detail space berdasarkan ID",
        "tags": [
          "Katalog Space"
        ]
      }
    },
    "/api/diskon/active": {
      "get": {
        "operationId": "DiskonController_aktif",
        "parameters": [
          {
            "name": "id_space",
            "required": false,
            "in": "query",
            "description": "Bila diisi, hanya promo milik pengelola space tersebut yang dikembalikan",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Daftar promo yang sedang aktif, dapat disaring per space",
        "tags": [
          "Katalog Diskon"
        ]
      }
    },
    "/api/diskon/check": {
      "post": {
        "operationId": "DiskonController_periksa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckPromoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Periksa validitas kode promo",
        "tags": [
          "Katalog Diskon"
        ]
      }
    },
    "/api/diskon/{id}": {
      "get": {
        "operationId": "DiskonController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Detail diskon berdasarkan ID",
        "tags": [
          "Katalog Diskon"
        ]
      }
    },
    "/api/reservasi": {
      "post": {
        "operationId": "ReservasiController_buat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservasiDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Buat pemesanan space baru",
        "tags": [
          "Reservasi Member"
        ]
      }
    },
    "/api/reservasi/my": {
      "get": {
        "operationId": "ReservasiController_milikSaya",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Seluruh pemesanan milik sendiri",
        "tags": [
          "Reservasi Member"
        ]
      }
    },
    "/api/reservasi/my/history": {
      "get": {
        "operationId": "ReservasiController_histori",
        "parameters": [
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 12,
              "example": 8,
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "example": 2026,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Histori pemesanan per bulan beserta rekapnya",
        "tags": [
          "Reservasi Member"
        ]
      }
    },
    "/api/reservasi/{id}/e-ticket": {
      "get": {
        "operationId": "ReservasiController_etiket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "E-Ticket reservasi beserta QR code check-in",
        "tags": [
          "Reservasi Member"
        ]
      }
    },
    "/api/reservasi/{id}": {
      "get": {
        "operationId": "ReservasiController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Detail reservasi berdasarkan ID",
        "tags": [
          "Reservasi Member"
        ]
      }
    },
    "/api/reservasi/{id}/cancel": {
      "patch": {
        "operationId": "ReservasiController_batalkan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Batalkan pemesanan sendiri",
        "tags": [
          "Reservasi Member"
        ]
      }
    }
  },
  "info": {
    "title": "Smart Space Booking API",
    "description": "REST API sistem reservasi coworking space untuk UKK RPL 2026/2027 Paket B. Mencakup autentikasi multi-role, katalog space, kode promo, reservasi, e-ticket ber-QR, check-in/check-out, serta rekapitulasi pendapatan bulanan.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Tempelkan access_token hasil login tanpa awalan \"Bearer\""
      }
    },
    "schemas": {
      "RegisterMemberDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "johndoe"
          },
          "password": {
            "type": "string",
            "example": "Secret123!"
          },
          "nama_member": {
            "type": "string",
            "example": "John Doe"
          },
          "instansi": {
            "type": "string",
            "example": "Universitas Indonesia / PT Maju Mundur"
          },
          "alamat": {
            "type": "string",
            "example": "Jl. Sudirman No. 123, Jakarta Selatan"
          },
          "telp": {
            "type": "string",
            "example": "081234567890"
          },
          "foto": {
            "type": "string",
            "example": "member_john.jpg",
            "description": "Nama file hasil unggah dari POST /api/upload/members"
          }
        },
        "required": [
          "username",
          "password",
          "nama_member",
          "instansi",
          "alamat",
          "telp"
        ]
      },
      "RegisterAdminSpaceDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "admin_space1"
          },
          "password": {
            "type": "string",
            "example": "Admin123!"
          },
          "nama_coworking": {
            "type": "string",
            "example": "Moklet Hub Coworking"
          },
          "nama_pemilik": {
            "type": "string",
            "example": "Ahmad Bidin"
          },
          "telp": {
            "type": "string",
            "example": "081298765432"
          },
          "alamat": {
            "type": "string",
            "example": "Jl. Danau Ranau No. 1, Malang"
          },
          "deskripsi": {
            "type": "string",
            "example": "Coworking space 3 lantai dengan WiFi 100 Mbps dan area parkir luas."
          },
          "foto": {
            "type": "string",
            "example": "moklet_hub.jpg"
          }
        },
        "required": [
          "username",
          "password",
          "nama_coworking",
          "nama_pemilik",
          "telp"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "johndoe",
            "description": "Username pengguna terdaftar"
          },
          "password": {
            "type": "string",
            "example": "Secret123!",
            "description": "Kata sandi akun pengguna"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
...
      "CreateSpaceDto": {
        "type": "object",
        "properties": {
          "nama_space": {
            "type": "string",
            "example": "Personal Desk Alpha 01"
          },
          "harga_per_jam": {
            "type": "number",
            "example": 25000,
            "description": "Tarif sewa per jam dalam Rupiah"
          },
          "tipe": {
            "type": "string",
            "enum": [
              "desk",
              "meeting_room",
              "private_office"
            ],
            "example": "desk"
          },
          "kapasitas": {
            "type": "number",
            "example": 1
          },
          "deskripsi": {
            "type": "string",
            "example": "WiFi 100Mbps, stopkontak, monitor 24 inch."
          },
          "foto": {
            "type": "string",
            "example": "desk_alpha_01.jpg"
          }
        },
        "required": [
          "nama_space",
          "harga_per_jam",
          "tipe",
          "kapasitas",
          "deskripsi"
        ]
      },
      "UpdateSpaceDto": {
        "type": "object",
        "properties": {
          "nama_space": {
            "type": "string",
            "example": "Personal Desk Alpha 01"
          },
          "harga_per_jam": {
            "type": "number",
            "example": 25000,
            "description": "Tarif sewa per jam dalam Rupiah"
          },
          "tipe": {
            "type": "string",
            "enum": [
              "desk",
              "meeting_room",
              "private_office"
            ],
            "example": "desk"
          },
          "kapasitas": {
            "type": "number",
            "example": 1
          },
          "deskripsi": {
            "type": "string",
            "example": "WiFi 100Mbps, stopkontak, monitor 24 inch."
          },
          "foto": {
            "type": "string",
            "example": "desk_alpha_01.jpg"
          }
        }
      },
...
      "UpdateDiskonDto": {
        "type": "object",
        "properties": {
          "nama_diskon": {
            "type": "string",
            "example": "PROMOAGUSTUS"
          },
          "persentase_diskon": {
            "type": "number",
            "example": 20,
            "minimum": 1,
            "maximum": 100
          },
          "tanggal_awal": {
            "type": "string",
            "example": "2026-08-01T00:00:00Z"
          },
          "tanggal_akhir": {
            "type": "string",
            "example": "2026-08-31T23:59:59Z"
          }
        }
      },
...
      "CheckPromoDto": {
        "type": "object",
        "properties": {
          "nama_diskon": {
            "type": "string",
            "example": "DISKONHEMAT20",
            "description": "Kode promo yang dimasukkan pengguna pada form checkout"
          },
          "id_space": {
            "type": "number",
            "example": 1,
            "description": "Space yang akan dipesan. Bila diisi, kepemilikan promo ikut diperiksa."
          }
        },
        "required": [
          "nama_diskon"
        ]
      },
      "CreateReservasiDto": {
        "type": "object",
        "properties": {
          "id_space": {
            "type": "number",
            "example": 1
          },
          "tanggal_reservasi": {
            "type": "string",
            "example": "2026-08-30"
          },
          "jam_mulai": {
            "type": "string",
            "example": "09:00"
          },
          "durasi_jam": {
            "type": "number",
            "example": 3,
            "minimum": 1
          },
          "id_diskon": {
            "type": "number",
            "example": 1,
            "description": "ID promo yang dipilih dari katalog diskon"
          },
          "kode_promo": {
            "type": "string",
            "example": "DISKONHEMAT20",
            "description": "Kode promo yang diketik manual; didahulukan bila keduanya dikirim"
          }
        },
        "required": [
          "id_space",
          "tanggal_reservasi",
          "jam_mulai",
          "durasi_jam"
        ]
      }
    }
  }
}